util: move dns_lease_key_lookup() to a separate module
